requirements.txt
requests>=2.28.0
click>=8.1.0
rich>=12.0.0
pandas>=1.5.0
numpy>=1.23.0
plotly>=5.13.0
kaleido>=0.2.1  # For static image export
python-dotenv>=0.21.0
